Juntando funiones en un solo archivo, dando primeros estilos, imagenes (logo propio y google)
