+++ src/index.html
@@ -22,23 +22,23 @@
 </head>
 <body>
   <div class="log">
-    <form class="sign-in">
-      <h2>Registrarse</h2>
-      <input id="usser" type="text" placeholder="Elige tu usuario">
-      <input id="email" type="email" placeholder="Ingresa e-mail">
-      <input id="password" type="password" placeholder="Ingresa contraseña">
-      <input id= "btn-sign-up" type="button" value="registrar">
-    </form>
-    <form class="log-in">
-      <h2>Iniciar Sesión</h2>
-      <input id="email2" type="email" placeholder="Ingresa e-mail">
-      <input id="password2" type="password" placeholder="Ingresa contraseña">
-      <input id= "btn-log-up" type="button" value="Iniciar">
-      <input id= "google-button" type="button" value="Google">
-    </form>  
-      <div id="contenido">
-        
-      </div>
+      <form class="register">
+        <h2>Registrarse</h2>
+        <input class= "space" id="usser" type="text" placeholder="Elige tu usuario">
+        <input class= "space" id="email" type="email" placeholder="Ingresa e-mail">
+        <input class= "space" id="password" type="password" placeholder="Ingresa contraseña">
+        <button class="button space" id= "btn-sign-up">Registrar</button>
+      </form>
+      <form class="register">
+        <h2>Iniciar Sesión</h2>
+        <input class= "space" id="email2" type="email" placeholder="Ingresa e-mail">
+        <input class= "space" id="password2" type="password" placeholder="Ingresa contraseña">
+        <button class="button space" id= "btn-log-up">Iniciar sesión</button>
+        <button class="button google space" id= "google-button"><img src="imgs/google-logo.png" alt="Google">Iniciar sesión con Google</button>
+      </form>  
+        <div id="contenido">
+          
+        </div>
     
 </div>
 <script type="module" src="main.js"></script>
