--- src/lib/logIn.js
@@ -1,12 +1,0 @@
-export const logIn = (email, password) => {
-
-    firebase.auth().signInWithEmailAndPassword(email, password)
-    .catch(function(error) {
-        // Handle Errors here.
-        let errorCode = error.code;
-        let errorMessage = error.message;
-        // ...
-      });
-  }
-
-
